Update wording in intro paragraph
It was weird to read the `I` form in the intro since this is a fork. I changed the `I` form to a more general `We` form and made it clear that this project is a fork including additional capabilites, such as temporal predicates. I hope that's alright with you guys. Happy to see that this is growing.